+++ src/cpp/server_core/include/server_core/http/SecureCookie.hpp
@@ -42,10 +42,13 @@ http::Cookie createSecureCookie(const std::string& name,
                                 const core::http::Request& request,
                                 const boost::posix_time::time_duration& validDuration,
                                 const std::string& path,
-                                bool secure);
+                                bool secure,
+                                bool iFrameEmbedding,
+                                bool legacyCookies);
 
 std::string readSecureCookie(const core::http::Request& request,
-                             const std::string& name);
+                             const std::string& name,
+                             bool iFrameLegacyCookies);
 
 std::string readSecureCookie(const std::string& signedCookieValue);
 
@@ -57,7 +60,10 @@ void set(const std::string& name,
          const boost::posix_time::time_duration& validDuration,
          const std::string& path,
          http::Response* pResponse,
-         bool secure);
+         bool secure,
+         bool iFrameEmbedding,
+         bool legacyCookies,
+         bool iFrameLegacyCookies);
 
 void set(const std::string& name,
          const std::string& value,
@@ -66,7 +72,10 @@ void set(const std::string& name,
          const boost::optional<boost::gregorian::days>& cookieExpiresDays,
          const std::string& path,
          http::Response* pResponse,
-         bool secure);
+         bool secure,
+         bool iFrameEmbedding,
+         bool legacyCookies,
+         bool iFrameLegacyCookies);
 
 void set(const std::string& name,
          const std::string& value,
@@ -75,13 +84,19 @@ void set(const std::string& name,
          const boost::optional<boost::posix_time::time_duration>& expiresFromNow,
          const std::string& path,
          http::Response* pResponse,
-         bool secure);
+         bool secure,
+         bool iFrameEmbedding,
+         bool legacyCookies,
+         bool iFrameLegacyCookies);
 
 void remove(const http::Request& request,
             const std::string& name,
             const std::string& path,
             core::http::Response* pResponse,
-            bool secure);
+            bool secure,
+            bool iFrameEmbedding,
+            bool legacyCookies,
+            bool iFrameLegacyCookies);
 
 // initialize with default secure cookie key file
 core::Error initialize();