+++ src/cpp/session/include/session/SessionConstants.hpp
@@ -83,6 +83,8 @@
 #define kPackageOutputInPackageFolder     "package-output-to-package-folder"
 
 #define kUseSecureCookiesSessionOption    "session-use-secure-cookies"
+#define kIFrameEmbeddingSessionOption     "session-iframe-embedding"
+#define kLegacyCookiesSessionOption       "legacy-cookies"
 
 // NOTE: literal versions of these are depended upon by the desktop/rsinverse
 // project so they should be updated there as well if they are changed